# ui/api-yaml/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /v1/resources/auth-user/update-password:
    post:
      tags:
        - auth-user-controller
      operationId: updatePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
        required: true
      responses:
        "200":
          description: OK
  /v1/resources/auth-user/update-login:
    post:
      tags:
        - auth-user-controller
      operationId: updateLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLoginDto'
        required: true
      responses:
        "200":
          description: OK
...
components:
  schemas:
    UpdatePasswordDto:
      required:
        - newPassword
        - oldPassword
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
    UpdateLoginDto:
      required:
        - newLogin
        - password
      type: object
      properties:
        newLogin:
          type: string
        password:
          type: string
...